chore: refactor read model configuration files into multiple files
