+++ apps/govquests-api/rails_app/app/read_models/action_tracking/on_action_execution_completed.rb
@@ -12,8 +12,6 @@ module ActionTracking
       else
         Rails.logger.error "Action execution not found: #{event.data[:execution_id]}"
       end
-    rescue ActiveRecord::RecordInvalid => e
-      Rails.logger.error "Failed to update action execution: #{e.message}"
     end
   end
 end
