+++ apps/govquests-api/rails_app/app/read_models/authentication/user_read_model.rb
@@ -1,0 +1,11 @@
+module Authentication
+class UserReadModel < ApplicationRecord
+    self.table_name = "users"
+
+    validates :user_id, presence: true, uniqueness: true
+    validates :address, presence: true, uniqueness: true
+    validates :user_type, presence: true, inclusion: {in: %w[delegate non_delegate]}
+
+    has_many :sessions, class_name: "Authentication::SessionReadModel", foreign_key: :user_id
+  end
+end
