+++ apps/govquests-api/rails_app/app/read_models/gamification/game_profile_read_model.rb
@@ -1,0 +1,7 @@
+module Gamification
+class GameProfileReadModel < ApplicationRecord
+    self.table_name = "user_game_profiles"
+
+    attribute :tier, :string
+  end
+end
