+++ apps/govquests-api/rails_app/app/read_models/gamification/leaderboard_read_model.rb
@@ -1,0 +1,12 @@
+module Gamification
+class LeaderboardReadModel < ApplicationRecord
+    self.table_name = "leaderboards"
+    self.primary_key = "leaderboard_id"
+
+    has_many :leaderboard_entries,
+      foreign_key: "leaderboard_id",
+      primary_key: "leaderboard_id",
+      class_name: "Gamification::LeaderboardEntryReadModel",
+      inverse_of: :leaderboard
+  end
+end
