+++ apps/govquests-api/rails_app/app/read_models/notifications/notification_template_read_model.rb
@@ -1,0 +1,10 @@
+module Notifications
+class NotificationTemplateReadModel < ApplicationRecord
+    self.table_name = "notification_templates"
+
+    validates :template_id, presence: true, uniqueness: true
+    validates :name, presence: true, uniqueness: true
+    validates :content, presence: true
+    validates :template_type, presence: true, inclusion: {in: ["email", "SMS", "push"]}
+  end
+end
