+++ apps/govquests-api/rails_app/app/read_models/questing/quest_action_read_model.rb
@@ -1,0 +1,8 @@
+module Questing
+class QuestActionReadModel < ApplicationRecord
+    self.table_name = "quest_actions"
+
+    belongs_to :quest, class_name: "Questing::QuestReadModel", foreign_key: "quest_id", primary_key: "quest_id"
+    belongs_to :action, class_name: "ActionTracking::ActionReadModel", foreign_key: "action_id", primary_key: "action_id"
+  end
+end
