+++ apps/govquests-api/rails_app/app/read_models/questing/read_model_configuration.rb
@@ -1,27 +1,10 @@
 module Questing
-  class QuestReadModel < ApplicationRecord
-    self.table_name = "quests"
-
-    has_many :quest_actions, -> { order(position: :asc) }, class_name: "Questing::QuestActionReadModel", foreign_key: "quest_id", primary_key: "quest_id"
-    has_many :actions, through: :quest_actions, source: :action, class_name: "ActionTracking::ActionReadModel"
-  end
-
-  class QuestActionReadModel < ApplicationRecord
-    self.table_name = "quest_actions"
-
-    belongs_to :quest, class_name: "Questing::QuestReadModel", foreign_key: "quest_id", primary_key: "quest_id"
-    belongs_to :action, class_name: "ActionTracking::ActionReadModel", foreign_key: "action_id", primary_key: "action_id"
-  end
-
-  class UserQuestReadModel < ApplicationRecord
-    self.table_name = "user_quests"
-  end
-
   class ReadModelConfiguration
     def call(event_store)
       event_store.subscribe(OnQuestCreated, to: [Questing::QuestCreated])
       event_store.subscribe(OnActionAssociatedWithQuest, to: [Questing::ActionAssociatedWithQuest])
       event_store.subscribe(OnQuestStarted, to: [Questing::QuestStarted])
+      event_store.subscribe(OnQuestCompleted, to: [Questing::QuestCompleted])
     end
   end
 end
