+++ apps/govquests-api/rails_app/app/read_models/questing/user_quest_read_model.rb
@@ -1,0 +1,5 @@
+module Questing
+class UserQuestReadModel < ApplicationRecord
+    self.table_name = "user_quests"
+  end
+end
