+++ apps/govquests-api/rails_app/app/read_models/rewarding/read_model_configuration.rb
@@ -1,16 +1,4 @@
 module Rewarding
-  class RewardReadModel < ApplicationRecord
-    self.table_name = "rewards"
-
-    def description
-      display_data["description"]
-    end
-
-    def image_url
-      display_data["image_url"]
-    end
-  end
-
   class ReadModelConfiguration
     def call(event_store)
       event_store.subscribe(OnRewardCreated, to: [Rewarding::RewardCreated])
