+++ apps/govquests-api/rails_app/app/read_models/rewarding/reward_read_model.rb
@@ -1,0 +1,8 @@
+module Rewarding
+class RewardReadModel < ApplicationRecord
+    self.table_name = "rewards"
+
+    def description
+      display_data["description"]
+    end
+end
