+++ apps/govquests-api/rails_app/config/application.rb
@@ -36,7 +36,7 @@ module GovquestsApi
     # not contain `.rb` files, or that should not be reloaded or eager loaded.
     # Common ones are `templates`, `generators`, or `middleware`, for example.
     config.autoload_lib(ignore: %w[assets tasks])
-
+    config.autoload_paths += Dir[Rails.root.join("../govquests")]
     # Configuration for the application, engines, and railties goes here.
     #
     # These settings can be overridden in specific environments using the files
